fix: choice type default value (#186)
