+++ src/test/java/fr/insee/pogues/test/JSONDeserializerTest.java
@@ -411,141 +411,6 @@ class JSONDeserializerTest {
     }
 
     @Test
-    void testChoiceTypeVariableResponses() throws JAXBException {
-        String json = """
-                {
-                   "Child": [
-                     {
-                       "type": "QuestionType",
-                       "Response": [
-                           {
-                                "choiceType": "VARIABLE"
-                           }
-                       ]
-                     }
-                   ]
-                 }
-                """;
-        JSONDeserializer deserializer = new JSONDeserializer();
-        Questionnaire questionnaire = deserializer.deserializeString(json);
-        assertEquals(ChoiceTypeEnum.VARIABLE, ((QuestionType) questionnaire.getChild().getFirst()).getResponse().getFirst().getChoiceType());
-    }
-
-    @Test
-    void testChoiceTypeSuggester() throws JAXBException {
-        String json = """
-            {
-               "Child": [
-                 {
-                   "type": "QuestionType",
-                   "Response": [
-                        {
-                            "choiceType": "SUGGESTER"
-                        }
-                   ]
-                 }
-               ]
-             }
-            """;
-
-        JSONDeserializer deserializer = new JSONDeserializer();
-        Questionnaire questionnaire = deserializer.deserializeString(json);
-        assertEquals(ChoiceTypeEnum.SUGGESTER, ((QuestionType) questionnaire.getChild().getFirst()).getResponse().getFirst().getChoiceType());
-    }
-
-    @Test
-    void testChoiceTypeDefaultValue() throws JAXBException {
-        String json = """
-                {
-                   "Child": [
-                     {
-                       "type": "QuestionType",
-                       "Response": [{}]
-                     }
-                   ]
-                 }
-                """;
-        JSONDeserializer deserializer = new JSONDeserializer();
-        Questionnaire questionnaire = deserializer.deserializeString(json);
-        assertEquals(ChoiceTypeEnum.CODE_LIST, ((QuestionType) questionnaire.getChild().getFirst()).getResponse().getFirst().getChoiceType());
-    }
-
-    @Test
-    void testOptionFilter() throws JAXBException {
-        String json = """
-                {
-                   "Child": [
-                     {
-                         "type": "QuestionType",
-                         "OptionFilter": "nvl($AGE$, 0) > 18"
-                     }
-                   ]
-                 }
-                """;
-        JSONDeserializer deserializer = new JSONDeserializer();
-        Questionnaire questionnaire = deserializer.deserializeString(json);
-        assertEquals("nvl($AGE$, 0) > 18", ((QuestionType) questionnaire.getChild().getFirst()).getOptionFilter());
-    }
-
-    @Test
-    void testVariableReferenceInResponse() throws JAXBException {
-        String json = """
-                {
-                   "Child": [
-                     {
-                       "type": "QuestionType",
-                       "Response": {
-                               "VariableReference": "id-loop-variable"
-                           }
-                     }
-                   ]
-                 }
-                """;
-        JSONDeserializer deserializer = new JSONDeserializer();
-        Questionnaire questionnaire = deserializer.deserializeString(json);
-        assertEquals(
-                "id-loop-variable",
-                ((QuestionType) questionnaire.getChild().getFirst())
-                        .getResponse().getFirst().getVariableReference());
-    }
-
-    @Test
-    void testVariableResponsesFullConfiguration() throws JAXBException {
-        String json = """
-            {
-               "Child": [
-                 {
-                   "type": "QuestionType",
-                   "questionType": "SINGLE_CHOICE",
-                   "OptionFilter": "nvl($AGE$, 0) > 18",
-                   "Response": {
-                       "choiceType": "VARIABLE",
-                       "VariableReference": "id-loop-variable"
-                   }
-                 }
-               ]
-             }
-            """;
-
-        JSONDeserializer deserializer = new JSONDeserializer();
-        Questionnaire questionnaire = deserializer.deserializeString(json);
-
-        QuestionType question =
-                (QuestionType) questionnaire.getChild().getFirst();
-
-
-        assertEquals(QuestionTypeEnum.SINGLE_CHOICE, question.getQuestionType());
-        assertEquals(ChoiceTypeEnum.VARIABLE, question.getResponse().getFirst().getChoiceType());
-        assertEquals("nvl($AGE$, 0) > 18", question.getOptionFilter());
-        assertNotNull(question.getResponse());
-        assertFalse(question.getResponse().isEmpty());
-        assertEquals(
-                "id-loop-variable",
-                question.getResponse().getFirst().getVariableReference()
-        );
-    }
-
-    @Test
     void testDeserializeVariableReferenceInExternalVariable() throws JAXBException {
 
         String json = """
